+++ test/integration/test_b2_command_line.py
@@ -49,6 +49,8 @@ from b2sdk.v2 import (
     UNKNOWN_FILE_RETENTION_SETTING,
 )
 
+from b2sdk.v2.exception import BucketIdNotFound, FileNotPresent
+
 SSE_NONE = EncryptionSetting(mode=EncryptionMode.NONE,)
 SSE_B2_AES = EncryptionSetting(
     mode=EncryptionMode.SSE_B2,
@@ -318,13 +320,24 @@ class Api:
                             file_version_info.id_, file_version_info.file_name, LegalHold.OFF
                         )
                     print('Removing file version:', file_version_info.id_)
-                    self.api.delete_file_version(file_version_info.id_, file_version_info.file_name)
+                    try:
+                        self.api.delete_file_version(
+                            file_version_info.id_, file_version_info.file_name
+                        )
+                    except FileNotPresent:
+                        print(
+                            'It seems that file version %s has already been removed' %
+                            (file_version_info.id_,)
+                        )
 
                 if files_leftover:
                     print('Unable to remove bucket because some retained files remain')
                 else:
                     print('Removing bucket:', bucket.name)
-                    self.api.delete_bucket(bucket)
+                    try:
+                        self.api.delete_bucket(bucket)
+                    except BucketIdNotFound:
+                        print('It seems that bucket %s has already been removed' % (bucket.name,))
                 print()